slibai/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
requests
apscheduler
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
httpx
python-dotenv
psycopg2-binary
